wkb_translate_wkt {wk}	R Documentation

Deprecated functions

Description

These functions are deprecated and will be removed in a future version.

Usage

wkb_translate_wkt(wkb, ..., precision = 16, trim = TRUE)

wkb_translate_wkb(wkb, ..., endian = NA_integer_)

wkt_translate_wkt(wkt, ..., precision = 16, trim = TRUE)

wkt_translate_wkb(wkt, ..., endian = NA_integer_)

Arguments

wkb
A list() of raw() vectors, such as that returned by sf::st_as_binary().

...
Used to keep backward compatibility with previous versions of these functions.

precision
The rounding precision to use when writing (number of decimal places).

trim
Trim unnecessary zeroes in the output?

endian
Force the endian of the resulting WKB.

wkt
A character vector containing well-known text.

[Package wk version 0.9.5 Index]