grd_tile_template {wk}	R Documentation

Compute overview grid tile

Description

A useful workflow for raster data in a memory bounded environment is to chunk a grid into sections or tiles. These functions compute tiles suitable for such processing. Use grd_tile_summary() to generate statistics for level values to choose for your application.

Usage

grd_tile_template(grid, level)

grd_tile_summary(grid, levels = NULL)

Arguments

grid
A grd_xy(), grd_rct(), or other object implementing grd_*() methods.

level
An integer describing the overview level. This is related to the step value by a power of 2 (i.e., a level of 1 indicates a step of 2, a level of 2 indicates a step of 4, etc.).

levels
A vector of level values or NULL to use a sequence from 0 to the level that would result in a 1 x 1 grid.

Value

A grd()

Examples

grid <- grd_rct(volcano)
grd_tile_summary(grid)
grd_tile_template(grid, 3)

[Package wk version 0.9.5 Index]