plot.wk_grd_xy {wk}	R Documentation

Plot grid objects

Description

Plot grid objects

Usage

## S3 method for class 'wk_grd_xy'
plot(x, ...)

## S3 method for class 'wk_grd_rct'
plot(
  x,
  ...,
  image = NULL,
  interpolate = FALSE,
  oversample = 4,
  border = NA,
  asp = 1,
  bbox = NULL,
  xlab = "",
  ylab = "",
  add = FALSE
)

Arguments

x
A wkb() or wkt()

...
Passed to plotting functions for features: graphics::points() for point and multipoint geometries, graphics::lines() for linestring and multilinestring geometries, and graphics::polypath() for polygon and multipolygon geometries.

image
A raster or nativeRaster to pass to graphics::rasterImage(). use NULL to do a quick-and-dirty rescale of the data such that the low value is black and the high value is white.

interpolate
Use TRUE to perform interpolation between color values.

oversample
A scale on the number of pixels on the device to use for sampling estimation of large raster values. Use Inf to disable.

border
Color to use for polygon borders. Use NULL for the default and NA to skip plotting borders.

asp, xlab, ylab
Passed to graphics::plot()

bbox
The limits of the plot as a rct() or compatible object

add
Should a new plot be created, or should handleable be added to the existing plot?

Value

x, invisibly.

Examples

plot(grd_rct(volcano))
plot(grd_xy(volcano))

[Package wk version 0.9.5 Index]