wk-package {wk}	R Documentation

wk: Lightweight Well-Known Geometry Parsing

Description

Provides a minimal R and C++ API for parsing well-known binary and well-known text representation of geometries to and from R-native formats. Well-known binary is compact and fast to parse; well-known text is human-readable and is useful for writing tests. These formats are useful in R only if the information they contain can be accessed in R, for which high-performance functions are provided here.

Author(s)

Maintainer: Dewey Dunnington dewey@fishandwhistle.net (ORCID)

Authors:

Edzer Pebesma edzer.pebesma@uni-muenster.de (ORCID)

Other contributors:

Anthony North anthony.jl.north@gmail.com [contributor]

See Also

Useful links:

https://paleolimbot.github.io/wk/

https://github.com/paleolimbot/wk

Report bugs at https://github.com/paleolimbot/wk/issues

[Package wk version 0.9.5 Index]