wk_count {wk}	R Documentation

Count geometry components

Description

Counts the number of geometries, rings, and coordinates found within each feature. As opposed to wk_meta(), this handler will iterate over the entire geometry.

Usage

wk_count(handleable, ...)

## Default S3 method:
wk_count(handleable, ...)

wk_count_handler()

Arguments

handleable
A geometry vector (e.g., wkb(), wkt(), xy(), rct(), or sf::st_sfc()) for which wk_handle() is defined.

...
Passed to the wk_handle() method.

Value

A data.frame with one row for every feature encountered and columns:

n_geom: The number of geometries encountered, including the root geometry. Will be zero for a null feature.

n_ring: The number of rings encountered. Will be zero for a null feature.

n_coord: The number of coordinates encountered. Will be zero for a null feature.

Examples

wk_count(as_wkt("LINESTRING (0 0, 1 1)"))
wk_count(as_wkb("LINESTRING (0 0, 1 1)"))

[Package wk version 0.9.5 Index]