wk_crs {wk}	R Documentation

Set and get vector CRS

Description

The wk package doesn't operate on CRS objects, but does propagate them through subsetting and concatenation. A CRS object can be any R object, and x can be any object whose 'crs' attribute carries a CRS. These functions are S3 generics to keep them from being used on objects that do not use this system of CRS propagation.

Usage

wk_crs(x)

## S3 method for class 'wk_vctr'
wk_crs(x)

## S3 method for class 'wk_rcrd'
wk_crs(x)

wk_crs(x) <- value

wk_set_crs(x, crs)

wk_crs_output(...)

wk_is_geodesic_output(...)

Arguments

x, ...
Objects whose "crs" attribute is used to carry a CRS.

value
See crs.

crs
An object that can be interpreted as a CRS

[Package wk version 0.9.5 Index]