wk_crs_equal {wk}	R Documentation

Compare CRS objects

Description

The wk_crs_equal() function uses special S3 dispatch on wk_crs_equal_generic() to evaluate whether or not two CRS values can be considered equal. When implementing wk_crs_equal_generic(), every attempt should be made to make wk_crs_equal(x, y) and wk_crs_equal(y, x) return identically.

Usage

wk_crs_equal(x, y)

wk_crs_equal_generic(x, y, ...)

Arguments

x, y
Objects stored in the crs attribute of a vector.

...
Unused

Value

TRUE if x and y can be considered equal, FALSE otherwise.

[Package wk version 0.9.5 Index]