wk_example {wk}	R Documentation

Create example geometry objects

Description

Create example geometry objects

Usage

wk_example(which = "nc", crs = NA, geodesic = FALSE)

wk_example_wkt

Arguments

which
An example name. Valid example names are

"nc" (data derived from the sf package)

"point", "linestring", "polygon", "multipoint", "multilinestring", "multipolygon", "geometrycollection"

One of the above with the "_z", "_m", or "_zm" suffix.

crs
An object that can be interpreted as a CRS

geodesic
TRUE if edges must be interpolated as geodesics when coordinates are spherical, FALSE otherwise.

Format

An object of class list of length 29.

Value

A wkt() with the specified example.

Examples

wk_example("polygon")

[Package wk version 0.9.5 Index]