wk_is_geodesic {wk}	R Documentation

Set and get vector geodesic edge interpolation

Description

Set and get vector geodesic edge interpolation

Usage

wk_is_geodesic(x)

wk_set_geodesic(x, geodesic)

wk_is_geodesic(x) <- value

wk_geodesic_inherit()

Arguments

x
An R object that contains edges

geodesic
TRUE if edges must be interpolated as geodesics when coordinates are spherical, FALSE otherwise.

value
See geodesic.

Value

TRUE if edges must be interpolated as geodesics when coordinates are spherical, FALSE otherwise.

[Package wk version 0.9.5 Index]