wk_problems {wk}	R Documentation

Validate well-known binary and well-known text

Description

The problems handler returns a character vector of parse errors and can be used to validate input of any type for which wk_handle() is defined.

Usage

wk_problems(handleable, ...)

wk_problems_handler()

Arguments

handleable
A geometry vector (e.g., wkb(), wkt(), xy(), rct(), or sf::st_sfc()) for which wk_handle() is defined.

...
Passed to the wk_handle() method.

Value

A character vector of parsing errors. NA signifies that there was no parsing error.

Examples

wk_problems(new_wk_wkt(c("POINT EMTPY", "POINT (20 30)")))
wk_handle(
  new_wk_wkt(c("POINT EMTPY", "POINT (20 30)")),
  wk_problems_handler()
)

[Package wk version 0.9.5 Index]