wk_trans_affine {wk}	R Documentation

Affine transformer

Description

Affine transformer

Usage

wk_trans_affine(trans_matrix)

wk_affine_identity()

wk_affine_rotate(rotation_deg)

wk_affine_scale(scale_x = 1, scale_y = 1)

wk_affine_translate(dx = 0, dy = 0)

wk_affine_fit(src, dst)

wk_affine_rescale(rct_in, rct_out)

wk_affine_compose(...)

wk_affine_invert(x)

Arguments

trans_matrix
A 3x3 transformation matrix

rotation_deg
A rotation to apply in degrees counterclockwise.

scale_x, scale_y
Scale factor to apply in the x and y directions, respectively

dx, dy
Coordinate offsets in the x and y direction

src, dst
Point vectors of control points used to estimate the affine mapping (using base::qr.solve()).

rct_in, rct_out
The input and output bounds

...
Zero or more transforms in the order they should be applied.

x
A wk_trans_affine()

[Package wk version 0.9.5 Index]