wk_void {wk}	R Documentation

Do nothing

Description

This handler does nothing and returns NULL. It is useful for benchmarking readers and handlers and when using filters that have side-effects (e.g., wk_debug()). Note that this handler stops on the first parse error; to see a list of parse errors see the wk_problems() handler.

Usage

wk_void(handleable, ...)

wk_void_handler()

Arguments

handleable
A geometry vector (e.g., wkb(), wkt(), xy(), rct(), or sf::st_sfc()) for which wk_handle() is defined.

...
Passed to the wk_handle() method.

Value

NULL

Examples

wk_void(wkt("POINT (1 4)"))
wk_handle(wkt("POINT (1 4)"), wk_void_handler())

[Package wk version 0.9.5 Index]