Subset grd objects

The grd_subset() method handles the subsetting of a grd() in x-y space. Ordering of indices is not considered and logical indies are recycled silently along dimensions. The result of a grd_subset() is always a grd() of the same type whose relationship to x-y space has not changed.

grd_subset(object, y = NULL, x = NULL, bbox = NULL, ...)

Arguments

object
A grd()

y, x
Raw indices. These must be equally spaced if passed as numeric; if passed as logical they are recycled silently along each dimension.

bbox
A bounding box to use as a subset. This is used to calculate a suitable y and x index vector representing all cells that intersect the bbox. Cells that only touch bbox on the bottom and right are not included in the subset, meaning you can safely tile a regularly-spaced grid along object without double-counting cells.

...
Passed to subset methods

Value

A modified grd().

Examples

grid <- grd_rct(volcano)
grd_subset(grid, seq(2, 61, by = 4), seq(2, 87, by = 4))

#> <wk_grd_rct [15 x 15] => [-0.5 27.5 59.5 87.5]>
#> List of 3
#>  $ data      : num [1:15, 1:15] 101 105 109 113 121 125 127 118 112 109 ...
#>  $ bbox      : wk_rct[1:1] [-0.5 27.5 59.5 87.5]
#>  $ data_order: chr [1:2] "x" "y"
#>  - attr(*, "class")= chr [1:2] "wk_grd_rct" "wk_grd"