Write geometry vectors

Write geometry vectors

# S3 method for sfc
wk_writer(handleable, ...)

# S3 method for sf
wk_writer(handleable, ...)

sfc_writer()

wkb_writer(buffer_size = 2048L, endian = NA_integer_)

wkt_writer(precision = 16L, trim = TRUE)

wk_writer(handleable, ...)

# S3 method for wk_wkt
wk_writer(handleable, ..., precision = 16, trim = TRUE)

# S3 method for wk_wkb
wk_writer(handleable, ...)

# S3 method for wk_xy
wk_writer(handleable, ...)

xy_writer()

Arguments

handleable
A geometry vector (e.g., wkb(), wkt(), xy(), rct(), or sf::st_sfc()) for which wk_handle() is defined.

...
Passed to the writer constructor.

buffer_size
Control the initial buffer size used when writing WKB.

endian
Use 1 for little endian, 0 for big endian, or NA for system endian.

precision
If trim is TRUE, the total number of significant digits to keep for each result or the number of digits after the decimal place otherwise.

trim
Use FALSE to keep trailing zeroes after the decimal place.

Value

A wk_handler.